# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "RECA", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/projects/{project_id}/approvals": {
      "get": {
        "tags": ["approvals"],
        "summary": "List Project Approvals",
        "operationId": "approvals_list_project_approvals_get_api_v1_projects_project_id_approvals",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ApprovalStatus" },
                { "type": "null" }
              ],
              "title": "Status"
            }
          },
          {
            "name": "approval_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ApprovalType" },
                { "type": "null" }
              ],
              "title": "Approval Type"
            }
          },
          {
            "name": "target_object_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "maxLength": 100 },
                { "type": "null" }
              ],
              "title": "Target Object Type"
            }
          },
          {
            "name": "requested_by_actor_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/AuditActorType" },
                { "type": "null" }
              ],
              "title": "Requested By Actor Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalListEnvelope"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}": {
      "get": {
        "tags": ["approvals"],
        "summary": "Get Approval",
        "operationId": "approvals_get_approval_get_api_v1_approvals_approval_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/approve": {
      "post": {
        "tags": ["approvals"],
        "summary": "Approve",
        "operationId": "approvals_approve_post_api_v1_approvals_approval_id_approve",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/reject": {
      "post": {
        "tags": ["approvals"],
        "summary": "Reject",
        "operationId": "approvals_reject_post_api_v1_approvals_approval_id_reject",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ApprovalRejectRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/cancel": {
      "post": {
        "tags": ["approvals"],
        "summary": "Cancel",
        "operationId": "approvals_cancel_post_api_v1_approvals_approval_id_cancel",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApprovalEnvelope" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalDecisionPublic": {
        "properties": {
          "status": { "$ref": "#/components/schemas/ApprovalStatus" },
          "user_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "User Id"
          },
          "decided_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Decided At"
          },
          "reason": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": ["status", "user_id", "decided_at", "reason"],
        "title": "ApprovalDecisionPublic"
      },
      "ApprovalDecisionRequest": {
        "properties": {
          "decision_reason": {
            "anyOf": [
              { "type": "string", "maxLength": 2000 },
              { "type": "null" }
            ],
            "title": "Decision Reason"
          },
          "item_decisions": {
            "items": { "$ref": "#/components/schemas/ApprovalItemDecision" },
            "type": "array",
            "title": "Item Decisions"
          }
        },
        "type": "object",
        "title": "ApprovalDecisionRequest"
      },
      "ApprovalEnvelope": {
        "properties": {
          "data": { "$ref": "#/components/schemas/ApprovalPublic" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "type": "object",
        "required": ["data", "meta"],
        "title": "ApprovalEnvelope"
      },
      "ApprovalItemDecision": {
        "properties": {
          "item_type": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Item Type"
          },
          "item_id": { "type": "string", "format": "uuid", "title": "Item Id" },
          "decision": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Decision"
          },
          "reason": {
            "anyOf": [
              { "type": "string", "maxLength": 2000 },
              { "type": "null" }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": ["item_type", "item_id", "decision"],
        "title": "ApprovalItemDecision"
      },
      "ApprovalItemPublic": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "item_type": { "type": "string", "title": "Item Type" },
          "item_id": { "type": "string", "format": "uuid", "title": "Item Id" },
          "decision": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Decision"
          },
          "reason": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": ["id", "item_type", "item_id", "decision", "reason"],
        "title": "ApprovalItemPublic"
      },
      "ApprovalListEnvelope": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ApprovalPublic" },
            "type": "array",
            "title": "Data"
          },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "type": "object",
        "required": ["data", "pagination", "meta"],
        "title": "ApprovalListEnvelope"
      },
      "ApprovalPublic": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "approval_type": { "$ref": "#/components/schemas/ApprovalType" },
          "target_object_type": {
            "type": "string",
            "title": "Target Object Type"
          },
          "target_object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Object Id"
          },
          "requester": {
            "$ref": "#/components/schemas/ApprovalRequesterPublic"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "status": { "$ref": "#/components/schemas/ApprovalStatus" },
          "decision": {
            "anyOf": [
              { "$ref": "#/components/schemas/ApprovalDecisionPublic" },
              { "type": "null" }
            ]
          },
          "payload_hash": { "type": "string", "title": "Payload Hash" },
          "payload_snapshot": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Payload Snapshot"
          },
          "impact_summary": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Impact Summary"
          },
          "expires_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Expires At"
          },
          "supersedes_approval_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Supersedes Approval Id"
          },
          "items": {
            "items": { "$ref": "#/components/schemas/ApprovalItemPublic" },
            "type": "array",
            "title": "Items"
          },
          "allowed_actions": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Allowed Actions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "approval_type",
          "target_object_type",
          "target_object_id",
          "requester",
          "requested_at",
          "status",
          "decision",
          "payload_hash",
          "impact_summary",
          "expires_at",
          "supersedes_approval_id",
          "items",
          "allowed_actions",
          "created_at"
        ],
        "title": "ApprovalPublic"
      },
      "ApprovalRejectRequest": {
        "properties": {
          "decision_reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Decision Reason"
          },
          "item_decisions": {
            "items": { "$ref": "#/components/schemas/ApprovalItemDecision" },
            "type": "array",
            "title": "Item Decisions"
          }
        },
        "type": "object",
        "required": ["decision_reason"],
        "title": "ApprovalRejectRequest"
      },
      "ApprovalRequesterPublic": {
        "properties": {
          "type": { "$ref": "#/components/schemas/AuditActorType" },
          "id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Id"
          }
        },
        "type": "object",
        "required": ["type", "id"],
        "title": "ApprovalRequesterPublic"
      },
      "ApprovalStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "CANCELLED",
          "EXPIRED",
          "SUPERSEDED"
        ],
        "title": "ApprovalStatus"
      },
      "ApprovalType": {
        "type": "string",
        "enum": [
          "RESEARCH_QUESTION_CONFIRMATION",
          "LITERATURE_DECISION_CONFIRMATION",
          "LITERATURE_EXTRACTION_CONFIRMATION",
          "CLEANING_PLAN_APPROVAL",
          "VARIABLE_ROLE_CONFIRMATION",
          "ANALYSIS_PLAN_APPROVAL",
          "FIGURE_CONFIRMATION",
          "MANUSCRIPT_FIX_APPROVAL",
          "CLAIM_CONFIRMATION",
          "EXPORT_CONFIRMATION"
        ],
        "title": "ApprovalType"
      },
...
      "AuditActorType": {
        "type": "string",
        "enum": ["USER", "AGENT", "SYSTEM", "WORKER"],
        "title": "AuditActorType"
      },
...
      "ContractErrorDetail": {
        "properties": {
          "code": { "type": "string", "title": "Code" },
          "message": { "type": "string", "title": "Message" },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "field_errors": {
            "items": { "additionalProperties": true, "type": "object" },
            "type": "array",
            "title": "Field Errors"
          },
          "request_id": { "type": "string", "title": "Request Id" },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "suggested_action": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Suggested Action"
          }
        },
        "type": "object",
        "required": ["code", "message", "request_id"],
        "title": "ContractErrorDetail"
      },
      "ContractErrorResponse": {
        "properties": {
          "error": { "$ref": "#/components/schemas/ContractErrorDetail" }
        },
        "type": "object",
        "required": ["error"],
        "title": "ContractErrorResponse"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": { "type": "integer", "title": "Page" },
          "page_size": { "type": "integer", "title": "Page Size" },
          "total": { "type": "integer", "title": "Total" },
          "total_pages": { "type": "integer", "title": "Total Pages" },
          "has_next": { "type": "boolean", "title": "Has Next" },
          "has_previous": { "type": "boolean", "title": "Has Previous" }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": { "type": "string", "title": "Request Id" },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "idempotency_replayed": {
            "type": "boolean",
            "title": "Idempotency Replayed",
            "default": false
          }
        },
        "type": "object",
        "required": ["request_id"],
        "title": "ResponseMeta"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}